+++ android/samples/test-app/src/main/java/com/liferay/mobile/screens/testapp/WebContentDisplayActivity.java
@@ -43,17 +43,17 @@ public class WebContentDisplayActivity extends ThemeActivity implements WebConte
 
 	@Override
 	public WebContent onWebContentReceived(WebContent html) {
-		info("Web Content received!");
+		info(getString(R.string.webcontent_received_info));
 		return null;
 	}
 
 	@Override
 	public void onWebContentClicked(WebView.HitTestResult result, MotionEvent event) {
-		info("Web Content clicked!");
+		info(getString(R.string.webcontent_clicked_info));
 	}
 
 	@Override
 	public void error(Exception e, String userAction) {
-		error("Could not receive web content information", e);
+		error(getString(R.string.webcontent_error), e);
 	}
 }