+++ android/samples/test-app/src/main/java/com/liferay/mobile/screens/testapp/WebContentListActivity.java
@@ -29,7 +29,7 @@ public class WebContentListActivity extends ThemeActivity implements BaseListLis
 
 	@Override
 	public void onListPageFailed(int startRow, Exception e) {
-		error("Page request failed", e);
+		error(getString(R.string.page_error), e);
 	}
 
 	@Override
@@ -41,7 +41,7 @@ public class WebContentListActivity extends ThemeActivity implements BaseListLis
 
 	@Override
 	public void onListPageReceived(int startRow, int endRow, List<WebContent> entries, int rowCount) {
-		info("Row " + startRow + " received!");
+		info(rowCount + " " + getString(R.string.rows_received_info) + " " + startRow);
 	}
 
 	@Override