+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Possible categories for the changes: Updated, Fixed, Features, Deprecated
 ####	 Updated
 * Base interactor not handling exceptions outside its domain anymore
 * List users with AssetListScreenlet
+####   Features
+* Spanish translations
 
 ###iOS
 
@@ -20,3 +22,5 @@ Possible categories for the changes: Updated, Fixed, Features, Deprecated
 * Expose user object in Session context instead of raw attributtes
 * List users with AssetListScreenlet
 * Detailed error messages
+####   Features
+* Spanish translations